\name{IrishDirectorates-package}
\alias{IrishDirectorates-package}
\docType{package}
\title{
\packageTitle{IrishDirectorates}
}
\description{
\packageDescription{IrishDirectorates}
}
\author{
Riccardo Rastelli
Maintainer: Riccardo Rastelli <riccardoras@gmail.com>
}
\references{
Friel, N., Rastelli, R., Wyse, J. and Raftery, A.E. (2016) <DOI:10.1073/pnas.1606295113>
}
\keyword{ package }
