\name{IrishDirectoratesFit}
\alias{IrishDirectoratesFit}
\docType{data}
\title{
Fitted Dynamic Bipartite Latent Position Model.
}
\description{
Fitted Dynamic Bipartite Latent Position Model (dblpm) that serves as initialisation for the Metropolis-within-Gibbs algorithm
}
\usage{data("IrishDirectoratesFit")}
\format{
The list \code{IrishDirectoratesFit} has one element called \code{tail} which contains the values for each of the model parameters.
}
\references{
Friel, N., Rastelli, R., Wyse, J. and Raftery, A.E. (2016) <DOI:10.1073/pnas.1606295113>.
}
\keyword{datasets}
